add -T option, auto select flux surface radii, and add BGLRS+FLUKA option
This commit has three major updates to gevgen_atmo:

1. add -T option

This command line option allows you to specify the livetime to simulate in
seconds instead of requiring you to specify a fixed number of events.

2. auto select flux surface radii

When using a ROOT geometry, gevgen_atmo will automatically pick suitable radii
for the flux generation surface based on the bounding box of the ROOT geometry.

3. add BGLRS+FLUKA option

I also added the ability to read in a new flux file format that I created which
combines the 2D high energy BGLRS fluxes with the 1D lower energy FLUKA fluxes.