automatically select flux surface radii for ROOT geometries
This commit updates gevgen_atmo to automatically select the longitudinal and
transverse radii for the flux surface when using a ROOT geometry and if the two
radii are not specified on the command line. If this is the case, both are set
to the diagonal of the bounding box of the geometry.